+++ src/types.d.ts
@@ -1,0 +1,9 @@
+import { FastifyInstance, FastifyBaseLogger } from "fastify";
+
+type Tlogger = FastifyBaseLogger | undefined;
+
+type TpluginOption = {
+  dir: string | string[];
+};
+
+export { Tlogger, TpluginOption };
